Add code given in lecture 3.

+++ Matchismo/Model/Card.m
@@ -16,16 +16,14 @@
 
 #pragma mark - Card Matching
 
-- (int)match:(NSArray *)otherCards
+- (NSInteger)match:(NSArray *)otherCards
 {
-    int score = 0;
-
+    NSInteger score = 0;
     for (Card *card in otherCards) {
         if ([card.contents isEqualToString:self.contents]) {
             score = 1;
         }
     }
-
     return score;
 }
